cloud-dock/playground_batch_ingest/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "playground-batch-ingest"
version = "0.1.0"
description = "Batch ingestion service for processing CSV files from GCS events"
authors = ["Data Engineering Team"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
flask = "3.1.1"
gunicorn = "23.0.0"
jsonschema = "4.24.0"
google-cloud-pubsub = "2.29.1"
google-cloud-storage = "2.18.0"
google-cloud-secret-manager = "2.24.0"
pandas = "2.2.0"
pydantic = "2.9.2"
requests = "2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "8.4.0"
pytest-cov = "4.0.0"
black = "25.1.0"
flake8 = "7.0.0"
mypy = "1.8.0"
isort = "5.13.2"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--verbose --cov=src --cov-report=term-missing --cov-report=html"
